+++ app/src/main/java/woowacourse/shopping/ui/component/cart/ShoppingCartCheckBox.kt
@@ -1,0 +1,18 @@
+package woowacourse.shopping.ui.component.cart
+
+import androidx.compose.material3.Checkbox
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ShoppingCartCheckBox(
+    onCheckedChange: (Boolean) -> Unit,
+    modifier: Modifier = Modifier,
+    checked: Boolean = false,
+) {
+    Checkbox(
+        checked = checked,
+        onCheckedChange = onCheckedChange,
+        modifier = modifier,
+    )
+}
