+++ app/src/main/java/woowacourse/shopping/ui/component/cart/ShoppingCartTopBar.kt
@@ -1,0 +1,57 @@
+package woowacourse.shopping.ui.component.cart
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import woowacourse.shopping.R
+import woowacourse.shopping.ui.theme.AndroidShoppingTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ShoppingCartTopBar(
+    onBackClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    TopAppBar(
+        title = {
+            Text(
+                text = stringResource(R.string.cart_top_bar_title),
+            )
+        },
+        navigationIcon = {
+            IconButton(onClick = onBackClick) {
+                Image(
+                    painter = painterResource(R.drawable.back_icon),
+                    contentDescription = stringResource(R.string.close_detail_description),
+                    modifier = Modifier.size(16.dp),
+                )
+            }
+        },
+        colors =
+            TopAppBarDefaults.topAppBarColors(
+                containerColor = MaterialTheme.colorScheme.surfaceContainer,
+                titleContentColor = MaterialTheme.colorScheme.onSurface,
+            ),
+        modifier = modifier,
+    )
+}
+@Composable
+@Preview(showBackground = true)
+private fun ShoppingCartTopBarPreview() {
+    AndroidShoppingTheme {
+        ShoppingCartTopBar(
+            onBackClick = {},
+        )
+    }
+}
