+++ app/src/main/java/woowacourse/shopping/ui/component/productlist/ProductItem.kt
@@ -1,6 +1,6 @@
 @file:Suppress("FunctionName")
 
-package woowacourse.shopping.ui.component
+package woowacourse.shopping.ui.component.productlist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -18,10 +17,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import woowacourse.shopping.R
+import woowacourse.shopping.model.Price
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.ProductTitle
+import woowacourse.shopping.ui.theme.AndroidShoppingTheme
 import java.text.DecimalFormat
 
 @Composable
@@ -55,7 +58,7 @@ fun ProductItem(
                         .height(154.dp)
                         .background(MaterialTheme.colorScheme.surfaceContainer),
             )
-            AddQuantityButton(
+            ProductItemCartAction(
                 quantity = quantity,
                 onAddToCartClick = onAddToCartClick,
                 onQuantityPlusClick = onQuantityPlusClick,
@@ -69,51 +72,31 @@ fun ProductItem(
             maxLines = 1,
             overflow = TextOverflow.Ellipsis,
             color = MaterialTheme.colorScheme.onBackground,
-            modifier =
-                Modifier.padding(
-                    horizontal = 7.5.dp,
-                ),
+            modifier = Modifier.padding(horizontal = 7.5.dp),
         )
         Text(
             text = DecimalFormat(stringResource(R.string.price_format_pattern)).format(product.getPrice()),
             color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier =
-                Modifier.padding(
-                    horizontal = 7.5.dp,
-                ),
+            modifier = Modifier.padding(horizontal = 7.5.dp),
         )
     }
 }
 
 @Composable
-private fun AddQuantityButton(
-    quantity: Int,
-    onAddToCartClick: () -> Unit,
-    onQuantityPlusClick: () -> Unit,
-    onQuantityMinusClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    if (quantity == 0) {
-        Box(
-            modifier = modifier.fillMaxWidth(),
-            contentAlignment = Alignment.BottomEnd,
-        ) {
-            ShoppingCardAddBox(
-                onShoppingCartAddClick = onAddToCartClick,
-                modifier = Modifier.size(36.dp),
-            )
-        }
-    } else {
-        Box(
-            modifier = modifier.fillMaxWidth(),
-            contentAlignment = Alignment.BottomCenter,
-        ) {
-            ProductQuantityBox(
-                onQuantityPlusClick = onQuantityPlusClick,
-                onQuantityMinusClick = onQuantityMinusClick,
-                quantity = quantity,
-                modifier = Modifier,
-            )
-        }
+@Preview(showBackground = true)
+private fun ProductItemPreview() {
+    AndroidShoppingTheme {
+        ProductItem(
+            product = Product(
+                id = 1L,
+                title = ProductTitle("샘플 상품"),
+                price = Price(12_000),
+                imageUrl = "",
+            ),
+            quantity = 1,
+            onAddToCartClick = {},
+            onQuantityPlusClick = {},
+            onQuantityMinusClick = {},
+        )
     }
 }
