+++ app/src/main/java/woowacourse/shopping/ui/screen/ShoppingCartScreen.kt
@@ -2,54 +2,25 @@
 
 package woowacourse.shopping.ui.screen
 
-import android.icu.text.DecimalFormat
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
-import woowacourse.shopping.R
 import woowacourse.shopping.model.Price
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.ProductTitle
 import woowacourse.shopping.model.ShoppingCartItem
 import woowacourse.shopping.model.ShoppingItem
-import woowacourse.shopping.ui.component.CartSkeletonItem
-import woowacourse.shopping.ui.component.ProductQuantityBox
+import woowacourse.shopping.ui.component.cart.CartSkeletonItem
+import woowacourse.shopping.ui.component.cart.ShoppingCartTopBar
+import woowacourse.shopping.ui.component.cart.ShoppingCartItemCard
 import woowacourse.shopping.ui.state.ShoppingCartState
 import woowacourse.shopping.ui.theme.AndroidShoppingTheme
 
@@ -102,7 +73,7 @@ fun ShoppingCartScreen(
                     verticalArrangement = Arrangement.spacedBy(12.dp),
                 ) {
                     shoppingCartItems.forEach { shoppingCartItem ->
-                        ShoppingCartItems(
+                        ShoppingCartItemCard(
                             shoppingCartItem = shoppingCartItem,
                             selected = shoppingCartItem.getId() in state.selectedCartItemIds,
                             quantityPrice = getQuantityPrice(shoppingCartItem),
@@ -117,229 +88,6 @@ fun ShoppingCartScreen(
             }
         }
     }
-}
-
-@Composable
-private fun ShoppingCartItems(
-    shoppingCartItem: ShoppingCartItem,
-    selected: Boolean,
-    quantityPrice: Int,
-    onRemoveShoppingItemClick: (ShoppingCartItem) -> Unit,
-    onToggleShoppingItemSelectionClick: (Long, Boolean) -> Unit,
-    onIncreaseShoppingItemQuantityClick: (ShoppingCartItem) -> Unit,
-    onDecreaseShoppingItemQuantityClick: (ShoppingCartItem) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .clip(
-                    RoundedCornerShape(4.dp),
-                ).border(
-                    color = MaterialTheme.colorScheme.outline,
-                    width = 1.dp,
-                    shape = RoundedCornerShape(4.dp),
-                ).padding(12.dp),
-    ) {
-        val product = shoppingCartItem.product
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.Bottom,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier.weight(1f),
-            ) {
-                ShoppingCartCheckBox(
-                    checked = selected,
-                    onCheckedChange = { isChecked ->
-                        onToggleShoppingItemSelectionClick(shoppingCartItem.getId(), isChecked)
-                    },
-                )
-                Text(
-                    text = product.getTitle(),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    color = MaterialTheme.colorScheme.onBackground,
-                )
-            }
-            Image(
-                painter = painterResource(R.drawable.remove_icon),
-                contentDescription = stringResource(R.string.remove_item_description),
-                modifier =
-                    Modifier
-                        .size(16.dp)
-                        .clickable { onRemoveShoppingItemClick(shoppingCartItem) },
-            )
-        }
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.Bottom,
-        ) {
-            AsyncImage(
-                model = product.imageUrl,
-                contentDescription = stringResource(R.string.product_image_description),
-                contentScale = ContentScale.Crop,
-                modifier =
-                    Modifier
-                        .width(136.dp)
-                        .height(72.dp)
-                        .background(MaterialTheme.colorScheme.surfaceContainer),
-            )
-            Column(
-                modifier =
-                    Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-                horizontalAlignment = Alignment.End,
-            ) {
-                ProductQuantityBox(
-                    onQuantityPlusClick = { onIncreaseShoppingItemQuantityClick(shoppingCartItem) },
-                    onQuantityMinusClick = { onDecreaseShoppingItemQuantityClick(shoppingCartItem) },
-                    quantity = shoppingCartItem.getQuantity(),
-                    modifier = Modifier.padding(top = 8.dp),
-                )
-                Text(
-                    text =
-                        DecimalFormat(stringResource(R.string.price_format_pattern)).format(
-                            quantityPrice,
-                        ),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun OrderButton(
-    shoppingCartItems: List<ShoppingCartItem>,
-    selectedCartItemIds: Set<Long>,
-    shoppingCartSelectItemCount: Int,
-    onToggleShoppingItemSelectionClick: (List<Long>, Boolean) -> Unit,
-    onOrderButtonClick: (List<Long>) -> Unit,
-    checked: Boolean,
-    orderComplete: Boolean,
-    totalPrice: Int,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .height(88.dp)
-                .background(MaterialTheme.colorScheme.surfaceContainer),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        if (orderComplete) {
-            Column(
-                modifier = modifier,
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                ShoppingCartCheckBox(
-                    checked = checked,
-                    onCheckedChange = { isChecked ->
-                        onToggleShoppingItemSelectionClick(
-                            shoppingCartItems.map { shoppingCartItem -> shoppingCartItem.getId() },
-                            isChecked,
-                        )
-                    },
-                )
-                Text(
-                    text = "전체",
-                    color = MaterialTheme.colorScheme.onSurface,
-                )
-            }
-        }
-        Text(
-            text =
-                DecimalFormat(
-                    stringResource(R.string.price_format_pattern),
-                ).format(totalPrice),
-            modifier =
-                Modifier
-                    .weight(2f)
-                    .padding(end = 16.dp),
-            textAlign = TextAlign.End,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-
-        Button(
-            onClick = { onOrderButtonClick(selectedCartItemIds.toList()) },
-            modifier =
-                Modifier
-                    .weight(1.5f)
-                    .fillMaxHeight(),
-            shape = RectangleShape,
-            enabled = shoppingCartSelectItemCount > 0,
-        ) {
-            Text(text = "주문하기($shoppingCartSelectItemCount)")
-        }
-    }
-}
-
-@Composable
-private fun ShoppingCartCheckBox(
-    onCheckedChange: (Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-    checked: Boolean = false,
-) {
-    Checkbox(
-        checked = checked,
-        onCheckedChange = onCheckedChange,
-        modifier = modifier,
-    )
-}
-
-@Composable
-@Preview(showBackground = true)
-private fun ShoppingCartOrderButtonPreview() {
-    AndroidShoppingTheme {
-        OrderButton(
-            shoppingCartItems = emptyList(),
-            selectedCartItemIds = emptySet(),
-            shoppingCartSelectItemCount = 3,
-            totalPrice = 3400000,
-            checked = true,
-            orderComplete = true,
-            onToggleShoppingItemSelectionClick = { _, _ -> },
-            onOrderButtonClick = {},
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ShoppingCartTopBar(
-    onBackClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    TopAppBar(
-        title = {
-            Text(
-                text = stringResource(R.string.cart_top_bar_title),
-            )
-        },
-        navigationIcon = {
-            IconButton(onClick = onBackClick) {
-                Image(
-                    painter = painterResource(R.drawable.back_icon),
-                    contentDescription = stringResource(R.string.close_detail_description),
-                    modifier = Modifier.size(16.dp),
-                )
-            }
-        },
-        colors =
-            TopAppBarDefaults.topAppBarColors(
-                containerColor = MaterialTheme.colorScheme.surfaceContainer,
-                titleContentColor = MaterialTheme.colorScheme.onSurface,
-            ),
-        modifier = modifier,
-    )
 }
 
 @Composable
@@ -422,26 +170,4 @@ private fun ShoppingCartScreenPreview() {
             onDecreaseShoppingItemQuantityClick = {},
         )
     }
-}
-
-@Composable
-@Preview(showBackground = true)
-private fun ShoppingCartItemsPreview() {
-    ShoppingCartItems(
-        shoppingCartItem =
-            ShoppingCartItem(
-                id = 1,
-                shoppingItem =
-                    ShoppingItem(
-                        Product(1, ProductTitle("동원 스위트콘"), Price(99_800), ""),
-                        4,
-                    ),
-            ),
-        quantityPrice = 399_200,
-        onRemoveShoppingItemClick = {},
-        onToggleShoppingItemSelectionClick = { _, _ -> },
-        onIncreaseShoppingItemQuantityClick = {},
-        onDecreaseShoppingItemQuantityClick = {},
-        selected = true,
-    )
 }
